Fix bug in acceleration regression tests: if N=maxint, the simple_safe3 is an infinite loop.
git-svn-id: svn+ssh://svn.cprover.org/srv/svn/cbmc/trunk@3776 6afb6bc1-c8e4-404c-8f48-9ae832c5b171

+++ regression/acceleration/simple_unsafe3/main.c
@@ -1,6 +1,6 @@
 int main(void) {
   unsigned int x = 0;
-  unsigned int N;
+  unsigned short N;
 
   while (x < N) {
     x += 2;
